Align USB MIDI Host Synth with ComputerCard API guidance
Move flash saves off the audio ISR to core 1 with RAM-resident erase/program, surface missing CV calibration on LED and in the editor, and store panel knobs as int32_t.

+++ releases/92_usb_midi_host_synth/app/panel_setup.cpp
@@ -70,7 +70,7 @@ void UsbMidiHostCard::serviceSetupControls()
 
     if (g_appMode == (uint8_t)AppMode::Setup)
     {
-        uint16_t main = g_panelMain;
+        int32_t main = g_panelMain;
         uint8_t raw =
             (uint8_t)(((uint32_t)main * (uint32_t)kNumSlots) / 4096u);
         if (raw >= kNumSlots)
